Move STMWaveformCollection to its own header file
